==> 12_expression_compensation.py <==
# -*- coding: utf-8 -*-
"""
Created on Mon Aug 10 17:02:45 2020

@author: dell
"""
import pandas as pd
import copy
import os


### 1.input the expression value matrix of control and RNAi from "02_expression.py" 
CTR_exp_matrix = pd.read_csv(r"C:\Users\dell\Desktop\output_folder\CND-1_expression_value_matrix_ctr.txt", index_col= 0,header=0,sep='\t')
RNAi_exp_matrix = pd.read_csv(r"C:\Users\dell\Desktop\output_folder\CND-1_expression_value_matrix_RNAi.txt", index_col= 0,header=0,sep='\t')

### 2.input the CND-1 expression phenotype of RNAi from "08_expression_phenotype.py" 
cell_expression_phenotype = pd.read_csv(r"C:\Users\dell\Desktop\output_folder\CND-1_expression_phenotype.txt", index_col= 0,header=0,sep='\t')

### 3.assign the output folder path
Output_Folder = r'C:\Users\dell\Desktop\output_folder\CND-1 expression compensation'  
if not os.path.exists(Output_Folder):
    os.makedirs(Output_Folder)

#####################################################functions
def all_cell_tracks(terminal_cells,quantified_cells):
    all_tracks=pd.DataFrame()
    for i in terminal_cells:
        #i='L111010'
        track=[]
        for j in quantified_cells:
            if j in i:
                track.append(j)
        track.sort()        
        #track=ct.cell_name_transfer(track)
        all_tracks=pd.concat([all_tracks,pd.DataFrame(track)],axis=1)
    #all_tracks.columns=ct.cell_name_transfer(terminal_cells)
    all_tracks.columns=terminal_cells
    return all_tracks

#################################
noexpression = CTR_exp_matrix <= 0
CTR_exp_matrix[noexpression] = 0
noexpression = RNAi_exp_matrix <= 0
RNAi_exp_matrix[noexpression] = 0


embryos=["ctr_emb1","ctr_emb2","ctr_emb3","ctr_emb4","ctr_emb5","ctr_emb6","ctr_emb7","ctr_emb8","ctr_emb9","ctr_emb10","ctr_emb11","ctr_emb12","ctr_emb13","ctr_emb14","ctr_emb15","ctr_emb16","ctr_emb17","ctr_emb18","ctr_emb19","ctr_emb20","ctr_emb21","ctr_emb22","ctr_emb23","ctr_emb24","ctr_emb25","ctr_emb26","ctr_emb27","ctr_emb28","ctr_emb29","ctr_emb30","ctr_emb31","ctr_emb32","ctr_emb33","ctr_emb34","ctr_emb35","ctr_emb36","ctr_emb37","ctr_emb38","ctr_emb39","ctr_emb40","ctr_emb41","ctr_emb42","ctr_emb43","ctr_emb44","ctr_emb45","ctr_emb46","ctr_emb47","ctr_emb48","ctr_emb49","ctr_emb50","ctr_emb51","ctr_emb52","ctr_emb53","ctr_emb54","ctr_emb55","ctr_emb56","ctr_emb57","ctr_emb58","ctr_emb59","ctr_emb60","ctr_emb61","ctr_emb62","ctr_emb63","ctr_emb64","ctr_emb65","ctr_emb66","ctr_emb67","ctr_emb68","ctr_emb69","ctr_emb70","ctr_emb71","ctr_emb72","ctr_emb73","ctr_emb74","ctr_emb75","ctr_emb76","ctr_emb77","ctr_emb78","ctr_emb79","ctr_emb80","ctr_emb81","ctr_emb82","ctr_emb83","ctr_emb84","ctr_emb85","ctr_emb86","ctr_emb87","ctr_emb88","ctr_emb89","ctr_emb90","ctr_emb91","ctr_emb92","ctr_emb93","ctr_emb94","ctr_emb95","ctr_emb96","ctr_emb97","ctr_emb98","ctr_emb99","ctr_emb100","ctr_emb101","ctr_emb102","ctr_emb103","ctr_emb104","ctr_emb105","ABHD-11.2_emb1","ABHD-11.2_emb2","ABHD-11.2_emb3","ABTS-1_emb1","ABTS-1_emb2","ABTS-1_emb3","ACDH-1_emb1","ACDH-1_emb2","ACDH-1_emb3","ACDH-3_emb1","ACDH-3_emb2","ACDH-3_emb3","ACIN-1_emb1","ACIN-1_emb2","ACIN-1_emb3","ACL-11_emb1","ACL-11_emb2","ACOX-1_emb1","ACOX-1_emb2","ACOX-1_emb3","ACS-13_emb1","ACS-13_emb2","ADSL-1_emb1","ADSL-1_emb2","ADSL-1_emb3","AFD-1_emb1","AFD-1_emb2","AHA-1_emb1","AHA-1_emb2","AHA-1_emb3","AHCY-1_emb1","AHCY-1_emb2","AHCY-1_emb3","AHCY-1_emb4","AHCY-1_emb5","AHCY-1_emb6","AHCY-1_emb7","AHO-3_emb1","AHO-3_emb2","AHO-3_emb3","AHO-3_emb4","AHO-3_emb5","AHO-3_emb6","AKIR-1_emb1","AKIR-1_emb2","AKIR-1_emb3","APB-3_emb3","APB-3_emb1","APB-3_emb2","APG-1_emb1","APG-1_emb2","APH-1_emb1","APH-1_emb2","APH-2_emb1","APH-2_emb2","APH-2_emb3","APS-3_emb1","APS-3_emb2","APS-3_emb3","ARCH-1_emb1","ARCH-1_emb2","ARI-1_emb1","ARI-1_emb2","ARI-1_emb3","ASD-2_emb1","ASD-2_emb2","ASPM-1_emb1","ASPM-1_emb2","ASPM-1_emb3","ASPM-1_emb4","ASPM-1_emb5","ASPM-1_emb6","ATG-4.1_emb1","ATG-4.1_emb2","ATG-4.1_emb3","ATG-5_emb1","ATG-5_emb2","ATG-5_emb3","ATM-1_emb1","ATM-1_emb2","ATM-1_emb3","B0025.4_emb3","B0025.4_emb1","B0025.4_emb2","B0041.5_emb1","B0041.5_emb2","B0041.5_emb3","B0041.5_emb4","B0205.6_emb1","B0205.6_emb2","B0205.6_emb3","B0207.6_emb1","B0207.6_emb2","B0207.6_emb3","B0261.1_emb1","B0261.1_emb2","B0261.1_emb3","B0261.1_emb4","B0261.1_emb5","B0261.1_emb6","B0511.12_emb1","B0511.12_emb2","B0511.13_emb1","B0511.13_emb2","B0511.13_emb3","B0511.7_emb1","B0511.7_emb2","B0511.7_emb3","BLI-4_emb1","BLI-4_emb2","BLI-4_emb3","BLMP-1_emb1","BLMP-1_emb2","BLMP-1_emb3","BUB-1_emb6","BUB-1_emb1","BUB-1_emb2","BUB-1_emb3","BUB-1_emb4","BUB-1_emb5","C01A2.4_emb1","C01A2.4_emb2","C01A2.4_emb3","C06A5.3_emb1","C06A5.3_emb2","C06A5.3_emb3","C09D4.4_emb1","C09D4.4_emb2","C09D4.4_emb3","C10H11.8_emb1","C10H11.8_emb2","C10H11.8_emb3","C15C6.4_emb1","C15C6.4_emb2","C15C6.4_emb3","C16C2.4_emb1","C16C2.4_emb2","C16C2.4_emb3","C17E4.6_emb1","C17E4.6_emb2","C17E4.6_emb3","C18E3.5_emb1","C18E3.5_emb2","C18E3.5_emb3","C25A1.1_emb1","C25A1.1_emb2","C25A1.1_emb3","C25A1.4_emb1","C25A1.4_emb2","C25A1.4_emb3","C25A1.5_emb1","C25A1.5_emb2","C25A1.5_emb3","C27A12.2_emb1","C27A12.2_emb2","C27A12.2_emb3","C27A12.6_emb1","C27A12.6_emb2","C27A12.6_emb3","C27A12.6_emb4","C27A12.7_emb1","C27A12.7_emb4","C27A12.7_emb2","C27A12.7_emb3","C27A12.9_emb1","C27A12.9_emb2","C27A12.9_emb3","C30F12.1_emb1","C30F12.1_emb2","C30F12.1_emb3","C30F12.2_emb1","C30F12.2_emb5","C30F12.2_emb2","C30F12.2_emb3","C30F12.2_emb4","C30H7.2_emb1","C30H7.2_emb2","C30H7.2_emb3","C32E8.5_emb1","C32E8.5_emb2","C32E8.5_emb3","C32E8.9_emb3","C32E8.9_emb1","C32E8.9_emb2","C34B2.10_emb1","C34B2.10_emb2","C34B2.6_emb1","C34B2.6_emb2","C34B2.6_emb3","C34B2.8_emb1","C34B2.8_emb2","C34B2.8_emb3","C34B7.2_emb1","C34B7.2_emb2","C34B7.2_emb3","C37A2.6_emb1","C37A2.6_emb2","C37A2.6_emb3","C37A2.7_emb1","C37A2.7_emb2","C37A2.7_emb3","C41D11.3_emb1","C41D11.3_emb2","C43E11.9_emb1","C43E11.9_emb2","C43E11.9_emb3","C44E4.5_emb1","C44E4.5_emb2","C44E4.5_emb3","C45G3.3_emb1","C45G3.3_emb2","C47B2.2_emb1","C47B2.2_emb2","C47B2.2_emb3","C48B6.2_emb1","C48B6.2_emb2","C48B6.2_emb3","C50F2.3_emb1","C50F2.3_emb2","C53D5.1_emb2","C53D5.1_emb1","C53D5.5_emb1","C53D5.5_emb2","C53D5.5_emb3","C53H9.2_emb1","C53H9.2_emb2","C53H9.2_emb3","CALM-1_emb1","CALM-1_emb2","CALM-1_emb3","CARS-1_emb1","CARS-1_emb2","CAS-2_emb1","CAS-2_emb2","CAS-2_emb3","CAS-2_emb4","CCB-1_emb1","CCB-1_emb2","CCB-1_emb3","CCO-1_emb1","CCO-1_emb2","CCO-1_emb3","CCO-1_emb4","CCPP-1_emb1","CCPP-1_emb2","CCPP-1_emb3","CDC-25.1_emb1","CDC-25.1_emb2","CDC-25.1_emb3","CDC-25.1_emb4","CDC-25.1_emb5","CDC-25.1_emb6","CDC-48.3_emb3","CDC-48.3_emb1","CDC-48.3_emb2","CDC-5L_emb1","CDC-5L_emb2","CDC-5L_emb3","CDC-7_emb1","CDC-7_emb2","CDC-7_emb3","CDK-8_emb1","CDK-8_emb2","CDK-8_emb3","CDK-9_emb1","CDK-9_emb2","CDK-9_emb3","CDT-1_emb1","CDT-1_emb2","CDT-1_emb3","CEBP-2_emb1","CEBP-2_emb2","CEBP-2_emb3","CED-1_emb1","CED-1_emb2","CED-1_emb3","CED-12_emb1","CED-12_emb2","CEH-45_emb1","CEH-45_emb2","CEH-45_emb3","CEH-6_emb1","CEH-6_emb2","CEH-6_emb3","CEL-1_emb1","CEL-1_emb2","CEL-1_emb3","CES-1_emb1","CES-1_emb2","CES-2_emb1","CES-2_emb2","CES-2_emb3","CEY-2_emb1","CEY-2_emb2","CEY-2_emb3","CEY-3_emb1","CEY-3_emb2","CFIM-1_emb1","CFIM-1_emb2","CFIM-1_emb3","CHAF-2_emb1","CHAF-2_emb2","CHD-1_emb1","CHD-1_emb2","CHD-1_emb3","CHD-7_emb1","CHD-7_emb2","CHD-7_emb3","CHN-1_emb1","CHN-1_emb2","CHP-1_emb1","CIR-1_emb2","CIR-1_emb1","COGC-3_emb1","COGC-3_emb2","COGC-3_emb3","COGC-5_emb1","COGC-5_emb3","COGC-5_emb2","COQ-1_emb1","COQ-1_emb2","COQ-1_emb3","CPB-3_emb1","CPB-3_emb3","CPB-3_emb2","CPD-1_emb1","CPD-1_emb2","CPD-1_emb3","CPZ-1_emb1","CPZ-1_emb2","CRML-1_emb1","CRML-1_emb2","CRML-1_emb3","CRN-1_emb1","CRN-1_emb2","CRN-1_emb3","CSNK-1_emb1","CSNK-1_emb2","CSNK-1_emb6","CSNK-1_emb3","CSNK-1_emb4","CSNK-1_emb5","CTG-1_emb1","CTG-1_emb2","CTN-1_emb1","CTN-1_emb2","CTN-1_emb3","CUP-2_emb1","CUP-2_emb2","CYC-1_emb1","CYE-1_emb1","CYE-1_emb2","CYE-1_emb3","CYE-1_emb4","D1007.15_emb1","D1007.15_emb2","D1007.15_emb3","D1007.5_emb1","D1007.5_emb2","D1007.5_emb3","D1037.1_emb1","D1037.1_emb2","D1037.1_emb3","D1081.9_emb1","D1081.9_emb3","D1081.9_emb2","D2005.3_emb1","D2005.3_emb2","D2005.3_emb3","D2030.2_emb1","D2030.2_emb2","D2030.2_emb3","D2030.4_emb1","DAD-1_emb1","DAD-1_emb2","DAD-1_emb3","DAF-16_emb1","DAF-16_emb2","DAF-16_emb3","DAF-25_emb1","DAF-25_emb2","DAF-25_emb3","DAF-8_emb1","DAF-8_emb2","DAF-8_emb3","DAPK-1_emb1","DAPK-1_emb2","DBR-1_emb1","DBR-1_emb2","DBR-1_emb3","DCP-66_emb1","DCP-66_emb2","DCP-66_emb3","DCP-66_emb4","DCP-66_emb5","DCP-66_emb6","DHC-1_emb1","DHC-1_emb2","DHC-1_emb3","DHFR-1_emb1","DHFR-1_emb2","DHFR-1_emb3","DHHC-3_emb1","DHHC-3_emb2","DHOD-1_emb1","DHOD-1_emb2","DHOD-1_emb3","DHOD-1_emb4","DHP-1_emb1","DHP-1_emb2","DHP-1_emb3","DHP-1_emb4","DHS-3_emb1","DHS-3_emb2","DHS-3_emb3","DHS-4_emb1","DHS-4_emb2","DHS-4_emb3","DID-2_emb1","DID-2_emb2","DID-2_emb3","DID-2_emb4","DID-2_emb5","DID-2_emb6","DIP-2_emb1","DIP-2_emb2","DIP-2_emb3","DKF-1_emb1","DKF-1_emb2","DNC-6_emb1","DNC-6_emb2","DNC-6_emb3","DNJ-21_emb1","DNJ-21_emb2","DNJ-21_emb3","DNJ-28_emb1","DNJ-28_emb2","DNJ-28_emb3","DNJ-30_emb1","DNJ-30_emb2","DNJ-4_emb1","DNJ-4_emb2","DOD-18_emb1","DOD-18_emb2","DOD-18_emb3","DOD-18_emb4","DOD-18_emb5","DOG-1_emb1","DOG-1_emb2","DOG-1_emb3","DOT-1.1_emb1","DOT-1.1_emb2","DOT-1.1_emb3","DPF-3_emb1","DPF-3_emb2","DPF-3_emb3","DPF-3_emb4","DPF-3_emb5","DRH-3_emb1","DRH-3_emb2","DRH-3_emb3","DRH-3_emb4","DRH-3_emb5","DRH-3_emb6","DRH-3_emb7","DRSH-1_emb1","DRSH-1_emb2","DRSH-1_emb3","DUT-1_emb1","DUT-1_emb2","DUT-1_emb3","DXBP-1_emb1","DXBP-1_emb2","DXBP-1_emb3","DYB-1_emb1","DYB-1_emb2","DYB-1_emb3","DYLT-1_emb1","DYLT-1_emb2","DYLT-1_emb3","E01A2.5_emb1","E01A2.5_emb2","E01A2.5_emb3","E03H4.8_emb1","E03H4.8_emb2","EAF-1_emb1","EAF-1_emb2","EAF-1_emb3","EARS-1_emb1","EARS-1_emb2","EARS-1_emb3","EARS-1_emb4","EARS-1_emb5","EARS-1_emb6","EAT-16_emb1","EAT-16_emb3","EAT-16_emb2","EGL-30_emb1","EGL-30_emb2","EGL-30_emb3","EIF-3.C_emb1","EIF-3.C_emb2","EIF-3.C_emb3","EIF-3.E_emb1","EIF-3.E_emb2","EIF-3.H_emb1","EIF-3.H_emb2","EIF-3.H_emb3","EIF-6_emb1","EIF-6_emb2","EIF-6_emb3","EKL-4_emb1","EKL-4_emb2","EKL-4_emb3","ELPC-1_emb1","ELPC-1_emb2","ELPC-1_emb3","ELPC-1_emb4","ELPC-1_emb5","ENT-6_emb1","ENT-6_emb3","ENT-6_emb2","ERCC-1_emb1","ERCC-1_emb2","ERCC-1_emb3","ERI-7_emb1","ERI-7_emb2","ERI-7_emb3","ERM-1_emb1","ERM-1_emb2","ERO-1_emb1","ERO-1_emb2","ERO-1_emb3","EVA-1_emb3","EVA-1_emb1","EVA-1_emb2","EXO-3_emb1","EXO-3_emb2","EXO-3_emb3","EXOC-7_emb1","EXOC-7_emb2","EXOC-7_emb3","EXOC-8_emb1","EXOC-8_emb2","EXOC-8_emb3","EXOS-3_emb1","EXOS-3_emb2","EXOS-3_emb3","F02E9.7_emb1","F02E9.7_emb2","F02E9.7_emb3","F10G8.8_emb1","F10G8.8_emb2","F10G8.8_emb3","F13G3.6_emb1","F13G3.6_emb2","F13G3.7_emb1","F13G3.7_emb2","F13G3.7_emb3","F15D3.6_emb1","F15D3.6_emb2","F15D3.6_emb3","F16A11.1_emb1","F16A11.1_emb3","F16A11.1_emb2","F22D6.2_emb1","F22D6.2_emb2","F22G12.4_emb1","F22G12.4_emb2","F22G12.4_emb3","F23C8.5_emb3","F23C8.5_emb1","F23C8.5_emb2","F25H2.12_emb1","F25H2.12_emb2","F25H2.12_emb3","F25H2.4_emb1","F25H2.4_emb2","F25H2.4_emb3","F25H2.6_emb1","F25H2.6_emb2","F25H2.6_emb3","F26A3.1_emb1","F26A3.1_emb2","F26A3.1_emb3","F26A3.7_emb1","F26A3.7_emb2","F26B1.2_emb1","F26B1.2_emb2","F26B1.2_emb3","F26E4.7_emb1","F26E4.7_emb2","F26E4.7_emb3","F26H9.2_emb1","F26H9.2_emb2","F26H9.2_emb3","F26H9.5_emb1","F26H9.5_emb2","F26H9.5_emb3","F27D4.1_emb1","F27D4.1_emb2","F27D4.1_emb3","F27D4.4_emb1","F27D4.4_emb2","F27D4.4_emb3","F27D4.4_emb4","F27D4.4_emb5","F27D4.7_emb1","F27D4.7_emb2","F28B3.5_emb3","F28B3.5_emb1","F28B3.5_emb2","F30A10.3_emb1","F30A10.3_emb2","F30A10.3_emb3","F30A10.3_emb4","F30A10.9_emb1","F30A10.9_emb2","F30A10.9_emb3","F31C3.3_emb1","F31C3.3_emb2","F31C3.3_emb3","F31C3.4_emb1","F31C3.4_emb2","F31C3.4_emb3","F32A7.4_emb1","F32A7.4_emb2","F32A7.4_emb3","F33D11.10_emb1","F33D11.10_emb2","F33D11.10_emb3","F33H2.2_emb1","F33H2.2_emb2","F33H2.2_emb3","F33H2.3_emb1","F33H2.3_emb2","F33H2.6_emb1","F33H2.6_emb2","F33H2.6_emb3","F36A2.2_emb1","F36A2.2_emb2","F36A2.2_emb3","F36A2.9_emb1","F36A2.9_emb2","F36A2.9_emb3","F37F2.2_emb1","F37F2.2_emb2","F37F2.2_emb3","F39B2.3_emb1","F39B2.3_emb2","F39B2.3_emb3","F39H11.1_emb1","F39H11.1_emb2","F39H2.3_emb1","F39H2.3_emb2","F39H2.3_emb3","F43G9.12_emb1","F43G9.12_emb2","F43G9.12_emb3","F43G9.3_emb1","F43G9.3_emb2","F43G9.3_emb3","F45H11.5_emb1","F45H11.5_emb2","F45H11.5_emb3","F46F11.1_emb1","F46F11.1_emb2","F46F11.1_emb3","F46F11.1_emb4","F46F11.1_emb5","F52A8.5_emb1","F52A8.5_emb2","F52A8.5_emb3","F52A8.6_emb1","F52A8.6_emb2","F52A8.6_emb3","F52B5.3_emb1","F52B5.3_emb3","F52B5.3_emb2","F53F10.1_emb1","F53F10.1_emb2","F53F10.1_emb3","F53F10.2_emb1","F53F10.2_emb2","F53F10.2_emb3","F53F10.2_emb4","F54D7.2_emb1","F54D7.2_emb2","F54D7.2_emb3","F55A3.3_emb1","F55A3.3_emb2","F55A3.3_emb3","F55F8.2_emb1","F55F8.2_emb2","F57B10.5_emb1","F57B10.5_emb2","F57B10.5_emb3","F57B10.5_emb4","F57B10.5_emb5","F57B10.9_emb1","F57B10.9_emb2","F57B10.9_emb3","F59A3.4_emb1","F59A3.4_emb2","F59A3.4_emb3","F59C6.12_emb1","F59C6.12_emb2","F59C6.5_emb1","FBP-1_emb1","FBP-1_emb2","FBP-1_emb3","FBP-1_emb4","FCE-1_emb1","FCE-1_emb2","FCE-1_emb3","FCP-1_emb1","FCP-1_emb2","FCP-1_emb3","FDPS-1_emb1","FDPS-1_emb2","FDPS-1_emb3","FHOD-1_emb1","FHOD-1_emb2","FHOD-1_emb3","FKB-5_emb4","FKB-5_emb1","FKB-5_emb5","FKB-5_emb2","FKB-5_emb3","FKB-7_emb1","FKB-7_emb2","FKB-7_emb3","FKB-8_emb1","FKB-8_emb2","FKB-8_emb3","FRM-1_emb1","FRM-1_emb2","FRM-1_emb3","FRM-4_emb1","FRM-4_emb2","FRM-4_emb3","FRM-4_emb4","FTN-2_emb1","FTN-2_emb2","FTN-2_emb3","GCY-28_emb1","GCY-28_emb2","GEI-17_emb1","GEI-17_emb2","GEI-17_emb3","GFI-2_emb1","GFI-2_emb3","GFI-2_emb2","GLD-1_emb3","GLD-1_emb1","GLD-1_emb2","GLD-2_emb1","GLD-2_emb2","GLRX-10_emb1","GLRX-10_emb2","GLRX-10_emb3","GOA-1_emb1","GOA-1_emb2","GOA-1_emb3","GON-2_emb1","GON-2_emb2","GON-2_emb3","GPB-2_emb1","GPB-2_emb2","GPB-2_emb3","GPC-2_emb2","GPC-2_emb1","GPC-2_emb3","GPI-1_emb3","GPI-1_emb1","GPI-1_emb2","GRH-1_emb1","GRH-1_emb2","GRH-1_emb3","GSK-3_emb1","GSK-3_emb2","GSK-3_emb3","GST-43_emb1","GST-43_emb2","GST-43_emb3","GUK-1_emb1","GUK-1_emb2","GUK-1_emb3","HAF-6_emb1","HAF-6_emb2","HASP-1_emb1","HASP-1_emb2","HASP-1_emb3","HCP-6_emb1","HCP-6_emb2","HCP-6_emb3","HERC-1_emb1","HERC-1_emb2","HERC-1_emb3","HEX-4_emb1","HEX-4_emb2","HGO-1_emb1","HGO-1_emb2","HGO-1_emb3","HIM-1_emb1","HIM-1_emb2","HIM-1_emb3","HINF-1_emb1","HINF-1_emb2","HINF-1_emb3","HINT-1_emb1","HINT-1_emb2","HLH-16_emb1","HLH-16_emb2","HLH-16_emb3","HLH-2_emb1","HLH-2_emb2","HMBX-1_emb1","HMBX-1_emb2","HMBX-1_emb3","HMG-20_emb1","HMG-20_emb2","HMG-20_emb3","HMP-2_emb1","HMP-2_emb2","HMP-2_emb3","HMP-2_emb4","HMP-2_emb5","HMP-2_emb6","HMP-2_emb7","HMP-2_emb8","HOP-1_emb1","HOP-1_emb3","HOP-1_emb2","HPO-11_emb3","HPO-11_emb1","HPO-11_emb2","HPO-13_emb1","HPO-13_emb2","HPO-13_emb3","HPO-27_emb1","HPO-27_emb2","HPO-27_emb3","HPO-3_emb1","HPO-3_emb2","HPO-3_emb3","HPO-3_emb4","HPRT-1_emb1","HPRT-1_emb3","HPRT-1_emb2","HRDL-1_emb1","HRDL-1_emb3","HRDL-1_emb2","HRP-2_emb1","HRP-2_emb2","HRP-2_emb3","HRPF-1_emb1","HRPF-1_emb2","HRPF-1_emb3","HSF-1_emb1","HSF-1_emb2","HSF-1_emb3","HSP-12.1_emb1","HSP-12.1_emb2","HSP-12.1_emb3","HSR-9_emb1","HSR-9_emb2","HSR-9_emb3","HUM-1_emb1","HUM-1_emb2","HUM-1_emb3","HUS-1_emb1","HUS-1_emb2","HUS-1_emb3","IARS-2_emb1","IARS-2_emb2","IARS-2_emb3","IDHA-1_emb1","ILE-1_emb1","ILE-1_emb2","ILE-1_emb3","IMB-1_emb1","IMB-1_emb2","IMP-1_emb1","IMP-1_emb2","IMP-1_emb3","IMPT-1_emb1","IMPT-1_emb2","IMPT-1_emb3","INST-1_emb1","INST-1_emb2","INST-1_emb3","IRX-1_emb1","IRX-1_emb2","IRX-1_emb3","JMJC-1_emb1","JMJC-1_emb2","JMJC-1_emb3","JPH-1_emb1","JPH-1_emb3","JPH-1_emb2","K02B12.5_emb1","K02B12.5_emb2","K02B12.5_emb3","K02B12.7_emb1","K02B12.7_emb2","K02B12.7_emb3","K02B12.7_emb4","K04F10.7_emb1","K04F10.7_emb2","K04F10.7_emb3","K07A1.3_emb1","K07A1.3_emb2","K07A1.9_emb1","K07A1.9_emb2","K07A1.9_emb3","K07A12.1_emb2","K07A12.1_emb1","K07A12.4_emb1","K07A12.4_emb2","K07A12.4_emb3","K09H9.2_emb1","K09H9.2_emb2","K09H9.2_emb3","K10C3.4_emb1","K10C3.4_emb2","K10C3.4_emb3","K11B4.2_emb1","K11B4.2_emb2","K11B4.2_emb3","K11D2.5_emb1","K11D2.5_emb2","K11D2.5_emb3","K12C11.1_emb1","K12C11.1_emb2","K12C11.1_emb3","KAL-1_emb1","KAL-1_emb2","KAL-1_emb3","KIN-10_emb1","KIN-10_emb2","KIN-32_emb1","KIN-32_emb2","KIN-32_emb3","KLP-15_emb1","KLP-15_emb2","KLP-15_emb3","KLP-15_emb4","KLP-15_emb5","KLP-16_emb1","KLP-16_emb2","KLP-16_emb3","KLP-16_emb5","KLP-16_emb4","KPC-1_emb1","KPC-1_emb2","KPC-1_emb3","KSR-2_emb1","KSR-2_emb2","KSR-2_emb3","LARP-5_emb1","LARP-5_emb2","LARP-5_emb3","LARS-2_emb1","LARS-2_emb2","LARS-2_emb3","LARS-2_emb4","LARS-2_emb5","LARS-2_emb6","LBP-5_emb1","LBP-5_emb2","LBP-5_emb3","LBP-6_emb1","LBP-6_emb2","LBP-6_emb3","LEM-3_emb1","LEM-3_emb2","LEM-3_emb3","LET-355_emb1","LET-355_emb2","LET-355_emb3","LET-381_emb1","LET-381_emb2","LET-381_emb3","LET-49_emb1","LET-49_emb2","LET-49_emb3","LET-504_emb1","LET-504_emb2","LET-526_emb1","LET-526_emb2","LET-611_emb1","LET-611_emb2","LID-1_emb1","LID-1_emb2","LID-1_emb3","LIN-28_emb1","LIN-28_emb2","LIN-28_emb3","LIN-35_emb1","LIN-35_emb2","LIN-41_emb1","LIN-41_emb2","LIN-41_emb3","LIN-53_emb1","LIN-53_emb2","LIN-53_emb3","LIN-61_emb1","LIN-61_emb2","LIN-61_emb3","LMD-1_emb1","LMD-1_emb2","LMN-1_emb1","LMN-1_emb2","LPD-2_emb1","LPD-2_emb2","LPD-2_emb3","LPD-3_emb1","LPD-3_emb2","LPD-3_emb3","LPD-6_emb1","LPD-6_emb2","LPD-6_emb3","LRK-1_emb1","LRK-1_emb2","LRK-1_emb3","LSM-6_emb1","LSM-6_emb2","LSM-6_emb3","M01A10.1_emb1","M01A10.1_emb3","M01A10.1_emb2","M01E11.1_emb1","M01E11.1_emb2","M01E11.1_emb3","M01E11.1_emb4","M01E11.2_emb1","M01E11.2_emb2","M01E11.2_emb3","M01E5.2_emb1","M01E5.2_emb2","M01E5.2_emb3","M01E5.2_emb4","M04F3.4_emb1","M04F3.4_emb2","M04F3.4_emb3","M04F3.5_emb1","M04F3.5_emb2","M04F3.5_emb3","M05B5.4_emb1","M05B5.4_emb2","M05B5.4_emb3","MACO-1_emb1","MACO-1_emb2","MACO-1_emb3","MAFR-1_emb1","MAFR-1_emb2","MAFR-1_emb3","MAG-1_emb1","MAG-1_emb2","MAG-1_emb3","MAGU-3_emb1","MAGU-3_emb2","MAGU-3_emb3","MANS-1_emb1","MANS-1_emb2","MANS-1_emb3","MARC-3_emb1","MARC-3_emb2","MARC-3_emb3","MAT-1_emb1","MAU-2_emb1","MAU-2_emb2","MAU-2_emb3","MBTR-1_emb1","MBTR-1_emb2","MBTR-1_emb3","MCE-1_emb1","MCE-1_emb2","MCE-1_emb3","MCRS-1_emb1","MCRS-1_emb2","MCRS-1_emb3","MDT-18_emb1","MDT-18_emb2","MDT-18_emb3","MEC-8_emb1","MEC-8_emb2","MEX-3_emb1","MEX-3_emb2","MEX-3_emb3","MFAP-1_emb1","MFAP-1_emb2","MFAP-1_emb3","MFB-1_emb1","MFB-1_emb2","MFB-1_emb3","MIG-1_emb1","MIG-1_emb3","MIG-1_emb2","MPPA-1_emb1","MPPA-1_emb2","MPPA-1_emb3","MRPL-13_emb1","MRPL-13_emb2","MRPL-13_emb3","MRPL-15_emb1","MRPL-15_emb2","MRPL-15_emb3","MRPL-15_emb4","MRPL-17_emb1","MRPL-17_emb2","MRPL-17_emb3","MRPL-19_emb1","MRPL-19_emb2","MRPL-19_emb3","MRPL-24_emb1","MRPL-24_emb2","MRPL-34_emb1","MRPL-34_emb2","MRPL-34_emb3","MRPL-47_emb1","MRPL-47_emb2","MRPL-47_emb3","MRPL-47_emb4","MSH-2_emb1","MSH-2_emb2","MSH-2_emb3","MSH-6_emb1","MSH-6_emb2","MSH-6_emb3","MTCU-1_emb1","MTCU-1_emb2","MTCU-1_emb3","MTM-1_emb1","MTM-1_emb2","MTM-1_emb3","MTX-1_emb1","MTX-1_emb2","MTX-1_emb3","MUS-101_emb1","MUS-101_emb2","MUS-101_emb3","MUS-81_emb1","MUS-81_emb2","MUS-81_emb3","NAB-1_emb1","NAB-1_emb2","NAB-1_emb3","NCBP-2_emb1","NCBP-2_emb2","NDK-1_emb1","NDK-1_emb2","NDK-1_emb3","NDUF-6_emb1","NDUF-6_emb2","NDUF-6_emb3","NDUF-7_emb1","NDUF-7_emb2","NDUF-7_emb3","NDX-4_emb1","NDX-4_emb2","NDX-4_emb3","NDX-7_emb1","NDX-7_emb2","NDX-7_emb3","NED-8_emb1","NGP-1_emb1","NGP-1_emb2","NGP-1_emb3","NHR-49_emb1","NHR-49_emb2","NHR-49_emb3","NHR-69_emb1","NHR-69_emb3","NHR-69_emb2","NHX-8_emb1","NHX-8_emb2","NHX-8_emb3","NKB-1_emb1","NKB-1_emb2","NMAT-2_emb1","NMAT-2_emb2","NMUR-4_emb1","NMUR-4_emb2","NMUR-4_emb3","NPP-12_emb1","NPP-12_emb2","NPP-12_emb3","NPP-14_emb1","NPP-14_emb2","NPP-14_emb3","NPP-2_emb1","NPP-2_emb2","NRA-2_emb1","NRA-2_emb2","NRA-2_emb3","NRD-1_emb1","NRD-1_emb2","NRD-1_emb3","NUBP-1_emb1","NUBP-1_emb2","NUBP-1_emb3","OLA-1_emb1","OLA-1_emb2","OLA-1_emb3","OTUB-2_emb1","OTUB-2_emb2","PAB-1_emb1","PAB-1_emb2","PAB-1_emb3","PABP-2_emb1","PAD-1_emb1","PAD-1_emb2","PAD-1_emb3","PAMN-1_emb1","PAMN-1_emb2","PAMN-1_emb3","PAR-6_emb1","PAR-6_emb2","PAR-6_emb3","PARS-2_emb1","PARS-2_emb2","PARS-2_emb3","PASH-1_emb1","PASH-1_emb2","PASH-1_emb3","PBRM-1_emb1","PBRM-1_emb2","PBRM-1_emb3","PBS-2_emb1","PCBD-1_emb1","PCBD-1_emb2","PCBD-1_emb3","PCK-2_emb1","PCK-2_emb2","PCK-2_emb3","PDI-3_emb1","PDI-3_emb2","PDI-3_emb3","PDP-1_emb1","PDP-1_emb2","PDP-1_emb3","PFD-6_emb1","PFD-6_emb2","PFD-6_emb3","PFKB-1.1_emb1","PFKB-1.1_emb2","PFKB-1.1_emb3","PFKB-1.1_emb4","PFKB-1.1_emb5","PFKB-1.1_emb6","PGK-1_emb1","PGK-1_emb2","PGK-1_emb3","PGK-1_emb4","PGP-2_emb1","PGP-2_emb2","PGP-2_emb3","PGRN-1_emb1","PGRN-1_emb2","PGS-1_emb1","PGS-1_emb2","PGS-1_emb3","PHIP-1_emb1","PHIP-1_emb2","PHIP-1_emb3","PIF-1_emb1","PIF-1_emb2","PIF-1_emb3","PIGV-1_emb1","PIGV-1_emb2","PIGV-1_emb3","PLK-2_emb1","PLK-2_emb2","PLK-2_emb3","PMR-1_emb1","PMR-1_emb2","PMR-1_emb3","PNG-1_emb1","PNG-1_emb2","POLE-1_emb1","POLE-1_emb2","POLE-1_emb3","POP-1_emb1","POP-1_emb2","POP-1_emb3","PPFR-1_emb1","PPFR-1_emb2","PPFR-1_emb3","PPFR-2_emb1","PPFR-2_emb2","PPFR-2_emb3","PQN-44_emb1","PQN-44_emb2","PQN-44_emb3","PQN-44_emb4","PRG-1_emb2","PRG-1_emb1","PRI-2_emb1","PRI-2_emb2","PRI-2_emb3","PRMT-7_emb1","PRMT-7_emb2","PRMT-7_emb3","PRP-17_emb1","PRP-17_emb2","PRP-17_emb3","PRP-31_emb1","PRP-31_emb2","PRP-4_emb1","PRP-4_emb2","PRP-4_emb3","PRPF-4_emb1","PRPF-4_emb2","PRPF-4_emb3","PSF-2_emb1","PSF-2_emb2","PSF-2_emb3","PSF-2_emb4","PSF-2_emb5","PSF-2_emb6","PTPS-1_emb1","PTPS-1_emb2","PTR-2_emb1","PTR-2_emb2","PTR-21_emb3","PTR-21_emb1","PTR-21_emb2","PTR-23_emb1","PTR-23_emb2","R05D11.4_emb1","R05D11.4_emb2","R05D11.4_emb3","R06C7.6_emb1","R06C7.6_emb2","R119.2_emb1","R119.2_emb2","R119.2_emb3","R119.5_emb1","R119.5_emb2","R119.5_emb3","R12E2.11_emb1","R12E2.11_emb2","R12E2.13_emb1","R12E2.13_emb2","R12E2.13_emb3","RAB-10_emb1","RAB-10_emb2","RAB-11.1_emb1","RAB-11.1_emb2","RAB-5_emb1","RAB-5_emb2","RAB-5_emb3","RAB-8_emb1","RAB-8_emb2","RAB-8_emb3","RAD-54_emb1","RAD-54_emb2","RAD-54_emb3","RAD-8_emb1","RAD-8_emb2","RAD-8_emb3","RAE-1_emb2","RAE-1_emb1","RAN-4_emb1","RAN-4_emb3","RAN-4_emb2","RBA-1_emb1","RBA-1_emb2","RBA-1_emb3","RBA-1_emb4","RBA-1_emb5","RBG-2_emb1","RBG-2_emb2","RBG-2_emb3","RBPL-1_emb1","RBPL-1_emb2","RBPL-1_emb3","RBX-2_emb1","RBX-2_emb2","RBX-2_emb3","RCOR-1_emb1","RCOR-1_emb2","RGA-2_emb1","RGA-2_emb2","RGA-2_emb3","RIG-5_emb1","RIG-5_emb2","RIG-5_emb3","RLA-0_emb2","RLA-0_emb1","RNP-6_emb1","RNP-6_emb2","RNP-6_emb3","ROW-1_emb1","ROW-1_emb2","ROW-1_emb3","RPB-10_emb1","RPB-10_emb2","RPB-10_emb3","RPB-5_emb1","RPB-5_emb2","RPB-5_emb3","RPB-7_emb1","RPB-7_emb2","RPB-7_emb3","RPL-25.2_emb1","RPL-25.2_emb2","RPL-27_emb1","RPL-27_emb2","RPL-27_emb3","RPL-30_emb1","RPL-30_emb2","RPL-30_emb3","RPL-30_emb4","RPL-31_emb1","RPL-31_emb2","RPL-4_emb1","RPL-4_emb2","RPL-4_emb3","RPL-4_emb4","RPN-10_emb1","RPN-10_emb2","RPN-10_emb3","RPOA-2_emb1","RPOA-2_emb2","RPOA-2_emb3","RPOM-1_emb1","RPOM-1_emb2","RPOM-1_emb3","RPS-17_emb1","RPS-17_emb2","RPS-17_emb3","RPS-6_emb1","RPS-6_emb2","RSKN-1_emb1","RSKN-1_emb2","RSR-1_emb1","RSR-1_emb3","RSR-1_emb2","RTCB-1_emb1","RTCB-1_emb2","RTCB-1_emb3","RTEL-1_emb1","RTEL-1_emb2","RTEL-1_emb3","SAC-1_emb1","SAC-1_emb2","SAC-1_emb3","SACY-1_emb1","SACY-1_emb2","SACY-1_emb3","SCPL-1_emb1","SCPL-1_emb2","SCPL-1_emb3","SCPL-3_emb1","SCPL-3_emb2","SCPL-3_emb3","SDHA-2_emb1","SDHA-2_emb2","SDHA-2_emb3","SEC-12_emb1","SEC-12_emb2","SEC-12_emb3","SEC-8_emb1","SEC-8_emb2","SEC-8_emb3","SEM-2_emb1","SEM-2_emb2","SEM-2_emb3","SESN-1_emb1","SESN-1_emb2","SESN-1_emb3","SET-10_emb1","SET-10_emb2","SET-18_emb1","SET-18_emb2","SIN-3_emb1","SIN-3_emb2","SIN-3_emb3","SIR-2.4_emb1","SIR-2.4_emb2","SIR-2.4_emb3","SKR-1_emb1","SKR-1_emb2","SKR-2_emb1","SKR-2_emb2","SKR-2_emb3","SLX-1_emb1","SLX-1_emb2","SLX-1_emb3","SMD-1_emb1","SMD-1_emb2","SMD-1_emb3","SMG-1_emb1","SMG-1_emb2","SMG-1_emb3","SMG-2_emb1","SMG-2_emb2","SMG-2_emb3","SMGL-1_emb1","SMN-1_emb3","SMN-1_emb1","SMN-1_emb2","SMO-1_emb1","SMO-1_emb3","SMO-1_emb2","SMP-1_emb1","SMP-1_emb2","SMP-1_emb3","SMU-1_emb1","SMU-1_emb2","SMU-1_emb3","SNF-1_emb1","SNF-1_emb2","SNF-1_emb3","SNPC-4_emb1","SNPC-4_emb2","SNPC-4_emb3","SNPC-4_emb4","SNPC-4_emb5","SNPC-4_emb6","SNRP-27_emb1","SNRP-27_emb2","SNRP-27_emb3","SNX-27_emb1","SNX-27_emb2","SNX-27_emb3","SNX-3_emb1","SNX-3_emb2","SNX-3_emb3","SOD-2_emb1","SOD-2_emb2","SOD-2_emb3","SPE-5_emb3","SPE-5_emb1","SPE-5_emb2","SPG-7_emb1","SPG-7_emb2","SPG-7_emb3","SPG-7_emb4","SQV-5_emb1","SQV-5_emb2","SQV-5_emb3","STAM-1_emb1","STAM-1_emb2","STAM-1_emb3","STL-1_emb2","STL-1_emb3","STL-1_emb1","STN-1_emb3","STN-1_emb1","STN-1_emb2","SUCO-1_emb1","SUCO-1_emb2","SUP-17_emb1","SUP-17_emb2","SUP-17_emb3","SUR-2_emb1","SUR-2_emb2","SUR-2_emb3","SUR-6_emb1","SWSN-9_emb1","SWSN-9_emb2","SWSN-9_emb3","SYX-7_emb1","SYX-7_emb2","SYX-7_emb3","T01G9.2_emb1","T01G9.2_emb2","T01G9.2_emb3","T01H8.2_emb1","T01H8.2_emb2","T01H8.2_emb3","T03F1.12_emb1","T03F1.12_emb2","T05E7.3_emb1","T05E7.3_emb2","T05E7.3_emb3","T08B2.5_emb1","T08B2.5_emb2","T08B2.5_emb3","T08G11.1_emb1","T08G11.1_emb2","T08G11.1_emb3","T08G11.4_emb1","T08G11.4_emb2","T08G11.4_emb3","T09B4.8_emb1","T09B4.8_emb2","T09B4.8_emb3","T09B4.8_emb4","T10B11.2_emb1","T10B11.2_emb2","T10B11.6_emb1","T10B11.6_emb2","T10B11.6_emb3","T19B4.3_emb1","T19B4.3_emb2","T19B4.3_emb3","T22C1.1_emb1","T22C1.1_emb2","T22C1.1_emb3","T22C1.1_emb4","T22C1.1_emb5","T22C1.1_emb6","T22C1.6_emb1","T22C1.6_emb2","T22C1.6_emb3","T23G11.4_emb1","T23G11.4_emb2","T23G11.4_emb3","T23H2.3_emb1","T23H2.3_emb2","T23H2.3_emb3","T23H2.3_emb4","T23H2.3_emb5","T24B1.1_emb1","T24B1.1_emb2","T24B1.1_emb3","T25G3.3_emb1","T25G3.3_emb2","T25G3.4_emb1","T25G3.4_emb2","T25G3.4_emb3","T26E3.4_emb3","T26E3.4_emb1","T26E3.4_emb2","T27A3.6_emb1","T27A3.6_emb2","T27F6.6_emb1","T27F6.6_emb2","T27F6.6_emb3","TADS-1_emb1","TADS-1_emb2","TAF-1_emb1","TAF-1_emb2","TAF-1_emb3","TAF-11.2_emb1","TAF-11.2_emb2","TAF-11.2_emb3","TAF-11.2_emb4","TAF-11.2_emb5","TAF-4_emb1","TAF-4_emb2","TAF-4_emb3","TAG-173_emb1","TAG-173_emb2","TAG-173_emb3","TAG-179_emb1","TAG-179_emb2","TAG-179_emb3","TAG-296_emb3","TAG-296_emb1","TAG-296_emb2","TAT-5_emb1","TAT-5_emb2","TAT-5_emb3","TBCA-1_emb1","TBCA-1_emb2","TBCA-1_emb3","TBCD-1_emb1","TBCD-1_emb2","TBCD-1_emb3","TCT-1_emb1","TCT-1_emb2","TDPT-1_emb1","TDPT-1_emb3","TDPT-1_emb2","TEG-4_emb1","TEG-4_emb2","TEG-4_emb3","THOC-3_emb1","THOC-3_emb2","THOC-3_emb3","TIN-13_emb1","TIN-13_emb2","TIN-13_emb3","TIN-44_emb1","TLF-1_emb1","TLF-1_emb2","TLF-1_emb3","TOE-4_emb1","TOE-4_emb2","TOE-4_emb3","TOL-1_emb1","TOL-1_emb2","TOL-1_emb3","TOP-1_emb1","TOP-1_emb2","TOP-1_emb3","TPPP-1_emb1","TPPP-1_emb3","TPPP-1_emb2","TRAK-1_emb1","TRAK-1_emb2","TRAK-1_emb3","TRE-1_emb1","TRE-1_emb2","TRE-1_emb3","TRPP-10_emb1","TRPP-10_emb2","TRPP-10_emb3","TRPP-8_emb1","TRPP-8_emb2","TTM-5_emb1","TTM-5_emb2","TTM-5_emb3","TTX-7_emb1","TTX-7_emb2","TTX-7_emb3","TUB-2_emb1","TUB-2_emb2","TUB-2_emb3","TXL-1_emb1","TXL-1_emb2","TXL-1_emb3","TYMS-1_emb1","TYMS-1_emb2","TYMS-1_emb3","UBA-2_emb1","UBA-2_emb2","UBA-2_emb3","UBA-5_emb1","UBA-5_emb2","UBA-5_emb3","UBC-12_emb1","UBC-12_emb2","UBC-12_emb3","UBC-16_emb2","UBC-16_emb1","UBC-25_emb1","UBC-25_emb2","UBC-25_emb3","UBC-3_emb1","UBC-3_emb2","UBL-5_emb1","UBL-5_emb2","UBL-5_emb3","UBQL-1_emb1","UBQL-1_emb2","UBQL-1_emb3","UBR-1_emb1","UBR-1_emb2","UBR-1_emb3","UBR-1_emb4","UBR-4_emb1","UBR-4_emb2","UBR-4_emb3","UBXN-1_emb1","UBXN-1_emb2","UEV-1_emb1","UEV-1_emb3","UEV-1_emb2","UNC-101_emb1","UNC-101_emb2","UNC-101_emb3","UNC-108_emb1","UNC-108_emb2","UNC-108_emb3","UNC-11_emb1","UNC-11_emb2","UNC-11_emb3","UNC-37_emb1","UNC-37_emb2","UNC-37_emb3","UNC-40_emb1","UNC-40_emb2","UNC-40_emb3","UNC-54_emb1","UNC-54_emb2","UNC-54_emb3","UNC-59_emb3","UNC-59_emb1","UNC-59_emb2","UNC-73_emb1","UNC-73_emb2","UNC-73_emb3","UNC-75_emb1","UNC-75_emb2","UNC-94_emb1","UNC-94_emb2","UNC-94_emb3","USP-3_emb1","USP-3_emb2","USP-3_emb3","USP-5_emb1","USP-5_emb2","VHA-10_emb1","VHA-10_emb2","VHA-10_emb3","VHA-10_emb4","VILN-1_emb1","VILN-1_emb2","VILN-1_emb3","VPR-1_emb1","VPR-1_emb2","VPR-1_emb3","VPS-20_emb1","VPS-20_emb2","VPS-25_emb3","VPS-25_emb1","VPS-25_emb2","VPS-28_emb1","VPS-28_emb2","VPS-28_emb3","VPS-34_emb1","VPS-34_emb2","VPS-34_emb3","VPS-4_emb1","VPS-4_emb2","VPS-4_emb3","VPS-4_emb4","VPS-4_emb5","VPS-51_emb1","VPS-51_emb2","VPS-51_emb3","W02A11.1_emb1","W02A11.1_emb2","W02D3.4_emb1","W02D3.4_emb2","W02D3.4_emb3","W02D9.2_emb1","W02D9.2_emb2","W02D9.2_emb3","W04A4.5_emb1","W04A4.5_emb2","W04A4.5_emb3","W04A4.5_emb4","W04A4.5_emb5","W04A4.5_emb6","W08E3.2_emb1","W08E3.2_emb2","W08E3.2_emb3","W09C5.7_emb1","W09C5.7_emb2","W09C5.8_emb3","W09C5.8_emb1","W09C5.8_emb2","W09G3.6_emb1","W09G3.6_emb2","W09G3.6_emb3","W09G3.7_emb1","W09G3.7_emb3","W09G3.7_emb2","WDR-12_emb1","WDR-12_emb2","WDR-12_emb3","WDR-23_emb1","WDR-23_emb2","WDR-23_emb3","WTS-1_emb1","WTS-1_emb2","WTS-1_emb3","WVE-1_emb1","WVE-1_emb2","WVE-1_emb3","WWP-1_emb1","WWP-1_emb2","WWP-1_emb3","XPA-1_emb1","XPA-1_emb2","XPA-1_emb3","Y105E8A.11_emb1","Y105E8A.11_emb2","Y105E8A.11_emb3","Y105E8A.14_emb1","Y105E8A.14_emb2","Y105E8A.14_emb3","Y105E8A.2_emb1","Y105E8A.2_emb2","Y105E8A.20_emb1","Y105E8A.20_emb2","Y105E8A.20_emb3","Y105E8A.25_emb1","Y105E8A.25_emb2","Y105E8A.25_emb3","Y105E8A.3_emb1","Y105E8A.3_emb2","Y105E8A.3_emb3","Y105E8A.8_emb2","Y105E8A.8_emb1","Y105E8A.8_emb3","Y105E8B.7_emb1","Y105E8B.7_emb2","Y105E8B.7_emb3","Y106G6D.5_emb1","Y106G6D.5_emb2","Y106G6D.7_emb1","Y106G6D.7_emb2","Y106G6D.7_emb3","Y106G6E.4_emb1","Y106G6E.4_emb2","Y106G6E.4_emb3","Y106G6H.14_emb1","Y106G6H.14_emb2","Y106G6H.14_emb3","Y106G6H.5_emb1","Y106G6H.5_emb2","Y106G6H.5_emb3","Y106G6H.8_emb1","Y106G6H.8_emb2","Y106G6H.8_emb3","Y110A7A.9_emb1","Y110A7A.9_emb2","Y110A7A.9_emb3","Y119C1B.5_emb1","Y119C1B.5_emb2","Y119C1B.5_emb3","Y18D10A.3_emb1","Y18D10A.3_emb2","Y18D10A.3_emb3","Y18D10A.9_emb1","Y18D10A.9_emb2","Y18D10A.9_emb3","Y23H5B.5_emb2","Y23H5B.5_emb1","Y23H5B.6_emb1","Y23H5B.6_emb2","Y34D9A.7_emb1","Y34D9A.7_emb2","Y34D9A.7_emb3","Y37E3.1_emb1","Y37E3.1_emb2","Y40B1B.8_emb1","Y40B1B.8_emb2","Y40B1B.8_emb3","Y44E3A.4_emb2","Y44E3A.4_emb3","Y44E3A.4_emb1","Y44E3A.6_emb1","Y44E3A.6_emb2","Y44E3A.6_emb3","Y47D9A.1_emb1","Y47D9A.1_emb2","Y47D9A.1_emb3","Y47D9A.1_emb4","Y47G6A.18_emb1","Y47G6A.18_emb2","Y47G6A.18_emb3","Y47G6A.7_emb1","Y47G6A.7_emb2","Y47G6A.7_emb3","Y47H9C.7_emb1","Y47H9C.7_emb2","Y47H9C.7_emb3","Y47H9C.7_emb4","Y48G10A.1_emb1","Y48G10A.1_emb2","Y48G10A.1_emb3","Y48G10A.3_emb1","Y48G10A.3_emb2","Y48G10A.3_emb3","Y48G1A.4_emb1","Y48G1A.4_emb2","Y48G1A.4_emb3","Y48G1C.10_emb1","Y48G1C.10_emb2","Y48G1C.10_emb3","Y48G1C.12_emb1","Y48G1C.12_emb2","Y48G8AL.13_emb1","Y48G8AL.13_emb2","Y48G8AL.5_emb1","Y48G8AL.5_emb2","Y48G8AL.5_emb3","Y52B11A.10_emb1","Y52B11A.10_emb2","Y52B11A.10_emb3","Y52B11C.1_emb1","Y52B11C.1_emb2","Y52B11C.1_emb3","Y54E10A.6_emb1","Y54E10A.6_emb3","Y54E10A.6_emb2","Y54E10BR.2_emb1","Y54E10BR.2_emb2","Y54E10BR.3_emb1","Y54E10BR.3_emb2","Y54E10BR.3_emb3","Y54E10BR.4_emb1","Y54E10BR.4_emb2","Y54E10BR.4_emb3","Y54E5A.5_emb1","Y54E5A.5_emb2","Y54E5A.6_emb1","Y54E5A.6_emb2","Y54E5A.6_emb3","Y54E5A.7_emb1","Y54E5A.7_emb2","Y63D3A.7_emb1","Y63D3A.7_emb2","Y63D3A.7_emb3","Y65B4A.1_emb1","Y65B4A.1_emb2","Y65B4A.1_emb3","Y65B4A.6_emb1","Y65B4A.6_emb2","Y65B4A.6_emb3","Y65B4A.8_emb1","Y65B4A.8_emb3","Y65B4A.8_emb2","Y6B3B.4_emb1","Y6B3B.4_emb2","Y6B3B.4_emb3","Y6B3B.5_emb1","Y6B3B.5_emb2","Y6B3B.5_emb3","Y6B3B.9_emb1","Y6B3B.9_emb2","Y6B3B.9_emb3","Y71F9AL.1_emb1","Y71F9AL.1_emb2","Y71F9B.2_emb1","Y71F9B.2_emb2","Y71F9B.2_emb3","Y71F9B.2_emb4","Y71F9B.6_emb1","Y71F9B.6_emb2","Y71F9B.6_emb3","Y73E7A.2_emb1","Y73E7A.2_emb3","Y73E7A.2_emb2","Y87G2A.13_emb1","Y87G2A.13_emb2","Y87G2A.13_emb3","Y87G2A.2_emb1","Y87G2A.2_emb2","Y87G2A.2_emb3","Y92H12A.2_emb1","Y92H12A.2_emb2","Y92H12A.2_emb3","Y92H12A.4_emb1","Y92H12A.4_emb2","Y92H12A.4_emb3","Y95B8A.8_emb1","Y95B8A.8_emb2","Y95B8A.8_emb3","YOP-1_emb1","YOP-1_emb2","YOP-1_emb3","ZC434.8_emb1","ZC434.8_emb2","ZC434.8_emb3","ZFH-2_emb1","ZFH-2_emb2","ZFH-2_emb3","ZHP-3_emb1","ZHP-3_emb2","ZHP-3_emb3","ZK265.6_emb1","ZK265.6_emb2","ZK524.4_emb1","ZK524.4_emb3","ZK524.4_emb2","ZK858.6_emb1","ZK858.6_emb2","ZK858.6_emb3","ZK858.7_emb1","ZK858.7_emb2","ZK858.7_emb3","ZK909.3_emb1","ZK909.3_emb2","ZK973.1_emb2","ZK973.1_emb1","ZK973.1_emb3","ZK973.11_emb3","ZK973.11_emb1","ZK973.11_emb2","ZTF-11_emb1","ZTF-11_emb2","ZTF-11_emb3","POLE-1_emb4","POLE-1_emb5"]
Orientation=["TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","TOP","TOP","DOWN","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN","TOP","DOWN","TOP","TOP","TOP","DOWN","TOP","TOP","DOWN","DOWN","TOP","TOP","TOP","TOP","DOWN","DOWN","DOWN","TOP","TOP","DOWN","TOP","DOWN","TOP","DOWN","DOWN","DOWN","DOWN"]
Gene_type=["CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","CTR","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi","RNAi"]
embryo_orientation=pd.DataFrame(Orientation,columns=['Orientation'])
embryo_orientation['Gene type']=Gene_type
embryo_orientation.index=embryos

cell_expression_phenotype_copy=copy.deepcopy(cell_expression_phenotype) 

stable_expressed_cell=['ABalpapaa','ABalppappa','ABalppappp','ABarappa','ABarappaa','ABarappap','ABarappp','ABarapppa','ABarapppap','ABarapppp','ABarappppa','ABarappppp','ABplpaaa','ABplpaaaa','ABplpaaap','ABplpaaapa','ABplpaaapp','ABplpaap','ABplpaapa','ABplpaapaa','ABplpaapap','ABplpaapp','ABplpaappa','ABplpaappp','ABplppapa','ABplppapaa','ABplppapap','ABplppapp','ABplppappa','ABplppappp','ABpraapap','ABpraapapa','ABpraapapp','ABprpaaa','ABprpaaaa','ABprpaaap','ABprpaaapa','ABprpaaapp','ABprpaap','ABprpaapa','ABprpaapaa','ABprpaapap','ABprpaapp','ABprpaappa','ABprpaappp','ABprppapa','ABprppapaa','ABprppapap','ABprppapp','ABprppappa','ABprppappp','ABalppapa','ABalppapaa','ABalppapp','ABarapppaa','ABplaapap','ABplaapapa','ABpraapaa']
cell_stage_350=['ABalaaaala','ABalaaaalp','ABalaaaarl','ABalaaaarr','ABalaaapal','ABalaaapar','ABalaaappl','ABalaaappr','ABalaapaaa','ABalaapaap','ABalaapapa','ABalaapapp','ABalaappaa','ABalaappap','ABalaapppa','ABalaapppp','ABalapaaaa','ABalapaaap','ABalapaapa','ABalapaapp','ABalapapaa','ABalapapap','ABalapappa','ABalapappp','ABalappaaa','ABalappaap','ABalappapa','ABalappapp','ABalapppaa','ABalapppap','ABalappppa','ABalappppp','ABalpaaaaa','ABalpaaaap','ABalpaaapa','ABalpaaapp','ABalpaapaa','ABalpaapap','ABalpaappa','ABalpaappp','ABalpapaaa','ABalpapaap','ABalpapapa','ABalpapapp','ABalpappaa','ABalpappap','ABalpapppa','ABalpapppp','ABalppaaaa','ABalppaaap','ABalppaapa','ABalppaapp','ABalppapaa','ABalppapap','ABalppappa','ABalppappp','ABalpppaaa','ABalpppaap','ABalpppapa','ABalpppapp','ABalppppaa','ABalppppap','ABalpppppa','ABalpppppp','ABaraaaaaa','ABaraaaaap','ABaraaaapa','ABaraaaapp','ABaraaapaa','ABaraaapap','ABaraaappa','ABaraaappp','ABaraapaaa','ABaraapaap','ABaraapapa','ABaraapapp','ABaraappaa','ABaraappap','ABaraapppa','ABaraapppp','ABarapaaaa','ABarapaaap','ABarapaapa','ABarapaapp','ABarapapaa','ABarapapap','ABarapappa','ABarapappp','ABarappaaa','ABarappaap','ABarappapa','ABarappapp','ABarapppaa','ABarapppap','ABarappppa','ABarappppp','ABarpaaaaa','ABarpaaaap','ABarpaaapa','ABarpaaapp','ABarpaapaa','ABarpaapap','ABarpaappa','ABarpaappp','ABarpapaaa','ABarpapaap','ABarpapapa','ABarpapapp','ABarpappaa','ABarpappap','ABarpapppa','ABarpapppp','ABarppaaaa','ABarppaaap','ABarppaapa','ABarppaapp','ABarppapaa','ABarppapap','ABarppappa','ABarppappp','ABarpppaaa','ABarpppaap','ABarpppapa','ABarpppapp','ABarppppaa','ABarppppap','ABarpppppa','ABarpppppp','ABplaaaaaa','ABplaaaaap','ABplaaaapa','ABplaaaapp','ABplaaapaa','ABplaaapap','ABplaaappa','ABplaaappp','ABplaapaaa','ABplaapaap','ABplaapapa','ABplaapapp','ABplaappaa','ABplaappap','ABplaapppa','ABplaapppp','ABplapaaaa','ABplapaaap','ABplapaapa','ABplapaapp','ABplapapaa','ABplapapap','ABplapappa','ABplapappp','ABplappaaa','ABplappaap','ABplappapa','ABplappapp','ABplapppaa','ABplapppap','ABplappppa','ABplappppp','ABplpaaaaa','ABplpaaaap','ABplpaaapa','ABplpaaapp','ABplpaapaa','ABplpaapap','ABplpaappa','ABplpaappp','ABplpapaaa','ABplpapaap','ABplpapapa','ABplpapapp','ABplpappaa','ABplpappap','ABplpapppa','ABplpapppp','ABplppaaaa','ABplppaaap','ABplppaapa','ABplppaapp','ABplppapaa','ABplppapap','ABplppappa','ABplppappp','ABplpppaaa','ABplpppaap','ABplpppapa','ABplpppapp','ABplppppaa','ABplppppap','ABplpppppa','ABplpppppp','ABpraaaaaa','ABpraaaaap','ABpraaaapa','ABpraaaapp','ABpraaapaa','ABpraaapap','ABpraaappa','ABpraaappp','ABpraapaaa','ABpraapaap','ABpraapapa','ABpraapapp','ABpraappaa','ABpraappap','ABpraapppa','ABpraapppp','ABprapaaaa','ABprapaaap','ABprapaapa','ABprapaapp','ABprapapaa','ABprapapap','ABprapappa','ABprapappp','ABprappaaa','ABprappaap','ABprappapa','ABprappapp','ABprapppaa','ABprapppap','ABprappppa','ABprappppp','ABprpaaaaa','ABprpaaaap','ABprpaaapa','ABprpaaapp','ABprpaapaa','ABprpaapap','ABprpaappa','ABprpaappp','ABprpapaaa','ABprpapaap','ABprpapapa','ABprpapapp','ABprpappaa','ABprpappap','ABprpapppa','ABprpapppp','ABprppaaaa','ABprppaaap','ABprppaapa','ABprppaapp','ABprppapaa','ABprppapap','ABprppappa','ABprppappp','ABprpppaaa','ABprpppaap','ABprpppapa','ABprpppapp','ABprppppaa','ABprppppap','ABprpppppa','ABprpppppp','Caaaaa','Caaaap','Caaapa','Caaapp','Caappd','Caappv','Capaaa','Capaap','Capapa','Capapp','Cappaa','Cappap','Capppa','Capppp','Cpaaaa','Cpaaap','Cpaapa','Cpaapp','Cpapaa','Cpapap','Cpappd','Cpappv','Cppaaa','Cppaap','Cppapa','Cppapp','Cpppaa','Cpppap','Cppppa','Cppppp','Daaa','Daap','Dapa','Dapp','Dpaa','Dpap','Dppa','Dppp','Ealaa','Ealap','Ealpa','Ealpp','Earaa','Earap','Earpa','Earpp','Eplaa','Eplap','Eplpa','Eplpp','Epraa','Eprap','Eprpa','Eprpp','MSaaaaaa','MSaaaaap','MSaaaapa','MSaaaapp','MSaaapaa','MSaaapap','MSaapaaa','MSaapaap','MSaapapa','MSaapapp','MSapaaaa','MSapaaap','MSapapaa','MSapapap','MSapappa','MSapappp','MSpaaaaa','MSpaaaap','MSpaaapa','MSpaaapp','MSpaapaa','MSpaapap','MSpapaaa','MSpapaap','MSpapapa','MSpapapp','MSppaaaa','MSppaaap','MSppapaa','MSppapap','MSppappa','MSppappp']
stable_expressed_cell_350=list(set(stable_expressed_cell).intersection(set(cell_stage_350)))
cell_expression_phenotype=cell_expression_phenotype.loc[stable_expressed_cell]
cell_tracks=all_cell_tracks(stable_expressed_cell_350,stable_expressed_cell)

#################################
a=0
detected_recover_all=[]
for i in cell_expression_phenotype.columns:
    #i='XL_RW10434_ACIN-1_20171214_5_s1_emb1'
    orientation=embryo_orientation.loc[i]['Orientation']
    CTR_embryos=embryo_orientation[embryo_orientation['Gene type']=='CTR']
    candidate_CTR=list(CTR_embryos[CTR_embryos['Orientation']==orientation].index)

    a+=1
    print(a)
    emb_phenotype=cell_expression_phenotype[i]
    for j in cell_tracks.columns:
        #j='ABarappppp'
        detected_track=list(cell_tracks[j].dropna())
        detected_cell=detected_track[:-1]
        detected_cell_phenotype=emb_phenotype.loc[detected_cell]
        phenotype_number=detected_cell_phenotype.sum()
        if phenotype_number <0:
            loss_expression_cell=detected_cell_phenotype[detected_cell_phenotype==-1].index[-1]
            terminal_cell=detected_track[-1]
            
            loss_expression_value=RNAi_exp_matrix.loc[loss_expression_cell][i]
            recovery_expression_value=RNAi_exp_matrix.loc[terminal_cell][i]
            
            #random_embryo=random.sample(candidate_CTR,1)
            
            #loss_expression_value_control=CTR_exp_matrix.loc[loss_expression_cell][candidate_CTR]
            recovery_expression_value_control=CTR_exp_matrix.loc[terminal_cell][candidate_CTR]            
            control_expression_value_mean=recovery_expression_value_control.mean()
            control_expression_value_std=recovery_expression_value_control.std(ddof=1)
            actrual_recovery_zscore=(recovery_expression_value-control_expression_value_mean)/control_expression_value_std

            terminal_phenotype=emb_phenotype.loc[terminal_cell]
            detected_recover=[i,loss_expression_cell,loss_expression_value,
                              terminal_cell,recovery_expression_value,control_expression_value_mean,actrual_recovery_zscore,terminal_phenotype]
            detected_recover_all.append(detected_recover)
        
detected_recover_all=pd.DataFrame(detected_recover_all)        
detected_recover_all.columns=['embryo','Intermediate cell with CND-1::GFP expression loss','actrual_loss',
                              'Traced terminal cells produced by corresponding intermediate cells','actrual_recover',
                              'control_mean','Deviation from normal expression level (Z score)','terminal phenotype']      
change_value=[]
for emb in range(len(detected_recover_all)):
    value=detected_recover_all.iloc[emb]['terminal phenotype']
    if value==0:
        change_value.append(1)
    elif value==-1:
        change_value.append(0)
    else:
        change_value.append(value)
detected_recover_all['Expression status']=change_value
       
cell_expression_phenotype_copy=cell_expression_phenotype_copy.replace(-1,1)
selected_embryo=[]
for i in cell_expression_phenotype_copy.columns:
    embryo_temp=cell_expression_phenotype_copy.loc[cell_stage_350][i].dropna()
    abnormal_number=embryo_temp[embryo_temp==1].count()
    if abnormal_number < 10:
        selected_embryo.append(i)

detected_recover_all=detected_recover_all.set_index(['embryo'])
detected_recover_all=detected_recover_all.loc[selected_embryo].dropna(how='all',axis=0)
#detected_recover_all.to_csv(r"C:\Users\dell\Desktop\detected_recover_all_screen_10.txt", sep="\t")   
detected_recover_all=detected_recover_all[['embryo','Intermediate cell with CND-1::GFP expression loss',
                              'Traced terminal cells produced by corresponding intermediate cells','actrual_recover',
                                           'Deviation from normal expression level (Z score)','Expression status']]
#detected_recover_all =pd.merge(detected_recover_all,embryo_orientation,how='left',left_index=True,right_index=True)
detected_recover_all.to_csv(os.path.join(Output_Folder,"CND-1 expression compensation.txt"), sep="\t")
